Fix matching of ISO specs when attaching previous info (#2039)
Builds that do not fetch the ISO catalog use previous info. The code that
associated previous info with the actual list of specs to process was only
matching on the spec's *computed* shortname. That computed shortname is a
temporary one for ISO specs, it gets changed right after that step when ISO
specs get processed. Code now also matches on the spec's URL if match on
shortname did not work.