Drop "delta" status for css-view-transitions-2 (#2119)
The Editor's Draft now contains the previous level, and the Web IDL interfaces
of Level 1 in particular. Curation in Webref fails accordingly.

Note: the /TR version hasn't been updated yet and is still to be viewed as a
delta spec over Level 1. We don't have a way to capture "delta in /TR,
non-delta in ED" though in the current data model though. This gives priority
to the `nightly` status because that's the reference one for data.