Add FIDO CTAP v2.2 and adjust shortnames of FIDO specs (#2126)
The list already had v2.1. This adds latest version v2.2.

The shortname was `fido-v2.1` which seems too generic given that the FIDO
Alliance publishes other specs that we might want to add in the future.
Shortnames now are derived from the filename that appears in the URL to disambiguate.

Fixes #2122.