Make sure alternateUrls are unique
When a network error occurs, build logic reuses previously published
information. Code that generates alternate URLs was not expecting that, and
duplicated the previous URL as a result.

Code now uses a `Set` to avoid any duplicate. Note it also sets the
`alternateUrls` directly for simplicity (and that behavior actually matches the
advertized one at the top of the file).