Throw when spec server returns a 404 response
This is meant to address cases such as #2134 and #2143 where a spec
"disappears" following a rename. Build code detected the problem but since we
don't have a good reporting mechanism in place, the code simply logged the
problem and moved on.

The code now throws when it receives a 404 response from the server, on the
grounds that data needs fixing.

The update disables the text on 429 errors. The problem is with the test
itself: the spec is loaded in Puppeteer, setting up a mock agent in the test
simply does not work. Test still passed because https://example.com/429 returns
a 404 and we had the same logic for 404 and 429 errors, but we've just been
lucky. Test now fails given that a 404 makes the code throw. To be fixed later
on!