Fully intercept network request when running tests
This fixes and re-enables the test that checks fallback logic when an HTTP
status 429 is received by fully intercepting HTTP requests that Puppeteer
sends, so that they can be mocked.

The use of an environment variable and the few test-related hardcoded lines in
the code's logic aren't fantastic, but that seemed the easiest way to capture
network requests for testing purpose.