Adjust series title test exception for TTML-IMSC
The title of the spec got updated in v1.3. We used to have an exception to the
rule for the test that checks alignment between the series title and the spec
title. The series tile for the `ttml-imsc` series has now been updated to match
that of v1.3, and the exception thus now needs to apply to v1.2 instead of 1.3.